Function 6 done.
